began refactoring controller routes. still deciding what can be kept and what needs to be removed alltogether.
