+++ controllers/api/index.js
@@ -1,8 +1,8 @@
 const router = require('express').Router();
-const userRoutes = require('./userRoutes');
-const projectRoutes = require('./projectRoutes');
+const loginRoutes = require('./loginRoutes');
+const calendarRoutes = require('./calendarRoutes');
 
-router.use('/users', userRoutes);
-router.use('/projects', projectRoutes);
+router.use('/login', loginRoutes);
+router.use('/calendar', calendarRoutes);
 
 module.exports = router;
